Run will not boot when game is already booted
Run calls Boot() which resets draw state and screen data. This is handy, because user does not have to call Boot() explicitly.

But sometimes user wants to run some code before game loop starts but after boot. This commit changes the behaviour of Run, which will not boot again if the game was booted before.